Put EbsConfig to a separate file
