+++ src/modules/aws-ebs/sub-modules/config/EbsConfig.ts
@@ -1,0 +1,6 @@
+export interface EbsConfig {
+  sourceAwsRegion: string;
+  targetAwsRegion: string;
+  targetAwsAccounts: string[];
+  transferAllSnapshots: boolean;
+}
