+++ src/modules/aws-ebs/sub-modules/snapshots/buildCopySnapshotDescription.ts
@@ -1,4 +1,4 @@
-import { EbsConfig } from "../loadEbsConfig";
+import { EbsConfig } from "../config/EbsConfig";
 
 export const buildCopySnapshotDescription = (config: EbsConfig, snapshotId: string): string => {
   return `[Ralph] Copied ${snapshotId} from ${config.sourceAwsRegion}`;
