+++ src/modules/aws-ebs/sub-modules/snapshots/isSnapshotInTargetRegion.ts
@@ -1,4 +1,4 @@
-import { EbsConfig } from "../loadEbsConfig";
+import { EbsConfig } from "../config/EbsConfig";
 
 export const isSnapshotInTargetRegion = (config: EbsConfig): boolean => {
   return config.sourceAwsRegion === config.targetAwsRegion;
